전체 도서조회 기능 결과코드

+++ controller/LikeController.js
@@ -1,36 +1,59 @@
+const jwt = require("jsonwebtoken");
 const conn = require("../mariadb"); //db모듈
 const { StatusCodes } = require("http-status-codes"); //status code모듈
+const ensureAuthorization = require("../auth");
 
 const addLike = (req, res) => {
     // 좋아요 추가
-    const { user_id } = req.body;
-    const { id } = req.params; //liked_book_id
+    // const { user_id } = req.body;
+    const book_id = req.params.id;
 
-    let sql = "INSERT INTO likes (user_id, liked_book_id) VALUES (?, ?)";
-    let values = [user_id, id];
-    conn.query(sql, values, (err, results) => {
-        if (err) {
-            console.log(err);
-            return res.status(StatusCodes.BAD_REQUEST).end();
-        }
-        return res.status(StatusCodes.OK).json(results);
-    });
+    let authorization = ensureAuthorization(req, res);
+    if (authorization instanceof jwt.TokenExpiredError) {
+        return res.status(StatusCodes.UNAUTHORIZED).json({
+            message: "로그인 세션이 만료되었습니다. 다시 로그인하세요.",
+        });
+    } else if (authorization instanceof jwt.JsonWebTokenError) {
+        return res.status(StatusCodes.BAD_REQUEST).json({
+            message: "잘못된 토큰입니다.",
+        });
+    } else {
+        let sql = "INSERT INTO likes (user_id, liked_book_id) VALUES (?, ?)";
+        let values = [authorization.id, book_id];
+        conn.query(sql, values, (err, results) => {
+            if (err) {
+                console.log(err);
+                return res.status(StatusCodes.BAD_REQUEST).end();
+            }
+            return res.status(StatusCodes.OK).json(results);
+        });
+    }
 };
 
 const removeLike = (req, res) => {
     // 좋아요 삭제
-    const { user_id } = req.body;
-    const { id } = req.params; //liked_book_id
 
-    let sql = "DELETE FROM likes WHERE user_id = ? AND liked_book_id =?";
-    let values = [user_id, id];
-    conn.query(sql, values, (err, results) => {
-        if (err) {
-            console.log(err);
-            return res.status(StatusCodes.BAD_REQUEST).end();
-        }
-        return res.status(StatusCodes.OK).json(results);
-    });
+    const book_id = req.params.id;
+    let authorization = ensureAuthorization(req, res);
+    if (authorization instanceof jwt.TokenExpiredError) {
+        return res.status(StatusCodes.UNAUTHORIZED).json({
+            message: "로그인 세션이 만료되었습니다. 다시 로그인하세요.",
+        });
+    } else if (authorization instanceof jwt.JsonWebTokenError) {
+        return res.status(StatusCodes.BAD_REQUEST).json({
+            message: "잘못된 토큰입니다.",
+        });
+    } else {
+        let sql = "DELETE FROM likes WHERE user_id = ? AND liked_book_id =?";
+        let values = [authorization.id, book_id];
+        conn.query(sql, values, (err, results) => {
+            if (err) {
+                console.log(err);
+                return res.status(StatusCodes.BAD_REQUEST).end();
+            }
+            return res.status(StatusCodes.OK).json(results);
+        });
+    }
 };
 
 module.exports = { addLike, removeLike };
